+++ src/main/java/com/getindata/connectors/http/BatchHttpStatusCodeValidationFailedException.java
@@ -1,0 +1,17 @@
+package com.getindata.connectors.http;
+
+import java.util.List;
+
+import lombok.Getter;
+
+import com.getindata.connectors.http.internal.sink.httpclient.HttpRequest;
+
+@Getter
+public class BatchHttpStatusCodeValidationFailedException extends Exception {
+    private final List<HttpRequest> failedRequests;
+
+    public BatchHttpStatusCodeValidationFailedException(String message, List<HttpRequest> failedRequests) {
+        super(message);
+        this.failedRequests = failedRequests;
+    }
+}