+++ src/test/java/com/getindata/connectors/http/internal/sink/HttpSinkConnectionTest.java
@@ -15,12 +15,14 @@ import com.github.tomakehurst.wiremock.stubbing.ServeEvent;
 import org.apache.flink.configuration.ConfigConstants;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.MetricOptions;
+import org.apache.flink.connector.base.DeliveryGuarantee;
 import org.apache.flink.metrics.Counter;
 import org.apache.flink.metrics.Metric;
 import org.apache.flink.metrics.MetricConfig;
 import org.apache.flink.metrics.MetricGroup;
 import org.apache.flink.metrics.reporter.MetricReporter;
 import org.apache.flink.metrics.reporter.MetricReporterFactory;
+import org.apache.flink.runtime.client.JobExecutionException;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.testutils.junit.extensions.ContextClassLoaderExtension;
 import org.junit.jupiter.api.AfterEach;
@@ -30,6 +32,7 @@ import org.junit.jupiter.api.extension.RegisterExtension;
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static com.github.tomakehurst.wiremock.stubbing.Scenario.STARTED;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.getindata.connectors.http.HttpSink;
@@ -189,18 +192,11 @@ public class HttpSinkConnectionTest {
     }
 
     @Test
-    public void testServerErrorConnection() throws Exception {
+    public void testNoRetryWithNoneConnection() throws Exception {
         wireMockServer.stubFor(any(urlPathEqualTo("/myendpoint"))
-                .withHeader("Content-Type", equalTo("application/json"))
                 .inScenario("Retry Scenario")
                 .whenScenarioStateIs(STARTED)
                 .willReturn(serverError())
-                .willSetStateTo("Cause Success"));
-        wireMockServer.stubFor(any(urlPathEqualTo("/myendpoint"))
-                .withHeader("Content-Type", equalTo("application/json"))
-                .inScenario("Retry Scenario")
-                .whenScenarioStateIs("Cause Success")
-                .willReturn(aResponse().withStatus(200))
                 .willSetStateTo("Cause Success"));
 
         var source = env.fromCollection(List.of(messages.get(0)));
@@ -210,17 +206,98 @@ public class HttpSinkConnectionTest {
                     (s, _context) ->
                     new HttpSinkRequestEntry("POST", s.getBytes(StandardCharsets.UTF_8)))
                 .setSinkHttpClientBuilder(JavaNetSinkHttpClient::new)
+                .setDeliveryGuarantee(DeliveryGuarantee.NONE)
+                .build();
+        source.sinkTo(httpSink);
+
+        env.execute("Http Sink none delivery guarantee retry");
+
+        assertEquals(1, SendErrorsTestReporterFactory.getCount());
+    }
+
+    @Test
+    public void testRetryableServerErrorConnection() throws Exception {
+        wireMockServer.stubFor(post(urlPathEqualTo("/myendpoint"))
+                .withRequestBody(equalTo("[{\"http-sink-id\":0}]"))
+                .inScenario("Retry Success Scenario")
+                .whenScenarioStateIs(STARTED)
+                .willReturn(serviceUnavailable())
+                .willSetStateTo("Cause Success"));
+
+        wireMockServer.stubFor(post(urlPathEqualTo("/myendpoint"))
+                .withRequestBody(equalTo("[{\"http-sink-id\":0}]"))
+                .inScenario("Retry Success Scenario")
+                .whenScenarioStateIs("Cause Success")
+                .willReturn(aResponse().withStatus(200).withBody("msg1"))
+                .willSetStateTo("Cause Success"));
+
+        var source = env.fromCollection(List.of(messages.get(0)));
+        var httpSink = HttpSink.<String>builder()
+                .setEndpointUrl("http://localhost:" + SERVER_PORT + "/myendpoint")
+                .setElementConverter(
+                    (s, _context) ->
+                        new HttpSinkRequestEntry("POST", s.getBytes(StandardCharsets.UTF_8)))
+                .setSinkHttpClientBuilder(JavaNetSinkHttpClient::new)
+                .setDeliveryGuarantee(DeliveryGuarantee.AT_LEAST_ONCE)
                 .build();
         source.sinkTo(httpSink);
         env.execute("Http Sink test failed connection");
 
         assertEquals(1, SendErrorsTestReporterFactory.getCount());
-        // TODO: reintroduce along with the retries
-        //  var postedRequests = wireMockServer
-        //  .findAll(postRequestedFor(urlPathEqualTo("/myendpoint")));
-        //  assertEquals(2, postedRequests.size());
-        //  assertEquals(postedRequests.get(0).getBodyAsString(),
-        //  postedRequests.get(1).getBodyAsString());
+
+        var postedRequests = wireMockServer
+            .findAll(postRequestedFor(urlPathEqualTo("/myendpoint")));
+        assertEquals(2, postedRequests.size());
+        assertEquals(postedRequests.get(0).getBodyAsString(), postedRequests.get(1).getBodyAsString());
+    }
+
+    @Test
+    public void testMixedRetryable() throws Exception {
+        wireMockServer.stubFor(post(urlPathEqualTo("/myendpoint"))
+                .withRequestBody(equalTo("[{\"http-sink-id\":0}]"))
+                .inScenario("Retry Success Scenario")
+                .whenScenarioStateIs(STARTED)
+                .willReturn(serviceUnavailable())
+                .willSetStateTo("Cause Success"));
+
+        wireMockServer.stubFor(post(urlPathEqualTo("/myendpoint"))
+                .withRequestBody(equalTo("[{\"http-sink-id\":0}]"))
+                .inScenario("Retry Success Scenario")
+                .whenScenarioStateIs("Cause Success")
+                .willReturn(aResponse().withStatus(200).withBody("msg1"))
+                .willSetStateTo("Cause Success"));
+
+        wireMockServer.stubFor(post(urlPathEqualTo("/myendpoint"))
+                .withRequestBody(equalTo("[{\"http-sink-id\":1}]"))
+                .inScenario("Retry Success Scenario")
+                .whenScenarioStateIs(STARTED)
+                .willReturn(aResponse().withStatus(200).withBody("msg2"))
+                .willSetStateTo(STARTED));
+
+        wireMockServer.stubFor(post(urlPathEqualTo("/myendpoint"))
+                .withRequestBody(equalTo("[{\"http-sink-id\":1}]"))
+                .inScenario("Retry Success Scenario")
+                .whenScenarioStateIs("Cause Success")
+                .willReturn(aResponse().withStatus(200).withBody("msg2"))
+                .willSetStateTo("Cause Success"));
+
+        var source = env.fromCollection(List.of(messages.get(0), messages.get(1)));
+        var httpSink = HttpSink.<String>builder()
+                .setEndpointUrl("http://localhost:" + SERVER_PORT + "/myendpoint")
+                .setElementConverter(
+                    (s, _context) ->
+                        new HttpSinkRequestEntry("POST", s.getBytes(StandardCharsets.UTF_8)))
+                .setSinkHttpClientBuilder(JavaNetSinkHttpClient::new)
+                .setDeliveryGuarantee(DeliveryGuarantee.AT_LEAST_ONCE)
+                .build();
+        source.sinkTo(httpSink);
+        env.execute("Http Sink test failed connection");
+
+        assertEquals(1, SendErrorsTestReporterFactory.getCount());
+
+        var postedRequests = wireMockServer
+                .findAll(postRequestedFor(urlPathEqualTo("/myendpoint")));
+        assertEquals(3, postedRequests.size());
     }
 
     @Test
@@ -249,7 +326,7 @@ public class HttpSinkConnectionTest {
                 .setSinkHttpClientBuilder(JavaNetSinkHttpClient::new)
                 .build();
         source.sinkTo(httpSink);
-        env.execute("Http Sink test failed connection");
+        assertThrows(JobExecutionException.class, () -> env.execute("Http Sink test failed connection"));
 
         assertEquals(1, SendErrorsTestReporterFactory.getCount());
         // var postedRequests = wireMockServer